#49 Implementaciónde detectarColision en la clase Objeto, y tratar colision de la clase Orbe

+++ app/src/main/java/grup05/pis2018/ub/edu/betweenopposites/Model/Posicion.kt
@@ -3,17 +3,17 @@ package grup05.pis2018.ub.edu.betweenopposites.Model
 /**
  * Clase que se encarga de tratar la posicion de los Objetos en pantalla
  */
-class Posicion(x: Int, y: Int) {
+class Posicion(x: Float, y: Float) {
 
     //Variables que contienen la posición
     //Las inicializo inicialmente a 0
-    var x: Int = 0
-    var y: Int = 0
+    var x: Float = 0f
+    var y: Float = 0f
 
     /**
      * Funcion para asignar una posicion (x,y)
      */
-    fun setPosicion(x: Int, y: Int) {
+    fun setPosicion(x: Float, y: Float) {
         this.x = x
         this.y = y
     }
